Minor adjustment and bug fixing

+++ umpoint/umpoint-admin/src/views/accommodation/accommodation-info.vue
@@ -152,13 +152,14 @@ import baseService from "@/service/baseService";
 import {useRoute} from "vue-router";
 import useView from "@/hooks/useView";
 import router from "@/router";
-import {ElMessage} from "element-plus";
+import {ElMessage, ElMessageBox} from "element-plus";
 import 'vue-cal/dist/vuecal.css';
 import VueCal from 'vue-cal';
 import UpdateBookingRule from "@/views/accommodation/booking-rule-add-or-update.vue";
 import UpdateClosure from "@/views/accommodation/closure-add-or-update.vue";
 import {formatDescription, getMondayAndSunday} from "@/utils/custom-utils";
 import {formatDateToDateTimeStr} from "@/utils/date";
+import {IObject} from "@/types/interface";
 
 const route = useRoute()
 const accommodation = ref();
@@ -289,15 +290,25 @@ const updateTimetable = (event: any) => {
 }
 
 const deleteHandle = () => {
-  baseService.delete("/accommodation/accommodation", [accommodation.value.id]).then((res) => {
-    ElMessage.success({
-      message: "Success",
-      duration: 500,
-      onClose: () => {
-        state.closeCurrentTab();
-      }
-    });
+  ElMessageBox.confirm(state.deleteMessage, "Hint", {
+    confirmButtonText: "Confirm",
+    cancelButtonText: "Cancel",
+    type: "warning"
   })
+    .then(() => {
+      baseService.delete("/accommodation/accommodation", [accommodation.value.id]).then((res) => {
+        ElMessage.success({
+          message: "Success",
+          duration: 500,
+          onClose: () => {
+            state.closeCurrentTab();
+          }
+        });
+      })
+    })
+    .catch(() => {
+      //
+    });
 }
 
 onMounted(() => {
